test and streamline registration failure handling
