+++ docs/reference/modules/registration.md
@@ -1,3 +1,9 @@
 # Image Registration Module
 
-::: merfish3danalysis.utils.registration
+## Affine registration and image warping
+
+::: merfish3danalysis.utils.multiview_registration
+
+## SOFIMA deformable registration
+
+::: merfish3danalysis.utils.sofima_registration
