+++ tests/test_sofima_deformable_registration.py
@@ -1,62 +1,7 @@
-import warnings
 from pathlib import Path
 
 import numpy as np
 import pytest
-
-
-def test_phase_shift_uses_maximum_overlap_alias_on_every_axis() -> None:
-    from merfish3danalysis.utils.multiview_registration import (
-        _maximum_overlap_phase_shift_px,
-    )
-
-    corrected = _maximum_overlap_phase_shift_px(
-        (62.1, -2047.2, 2046.75),
-        (63, 2048, 2048),
-    )
-    np.testing.assert_allclose(corrected, (-0.9, 0.8, -1.25), atol=1e-10)
-
-
-def test_register_pair_to_fixed_corrects_near_full_axis_aliases(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    cp = pytest.importorskip("cupy")
-    registration = pytest.importorskip("cucim.skimage.registration")
-
-    from merfish3danalysis.utils.multiview_registration import register_pair_to_fixed
-
-    try:
-        cp.cuda.runtime.getDeviceCount()
-    except cp.cuda.runtime.CUDARuntimeError:
-        pytest.skip("CUDA device is not available.")
-
-    def phase_cross_correlation(
-        fixed: object,
-        moving: object,
-        **kwargs: object,
-    ) -> tuple[object, float, float]:
-        del moving
-        assert kwargs["disambiguate"] is False
-        shift = (15.1, -15.2) if fixed.ndim == 2 else (-62.1, 0.0, 0.0)
-        return cp.asarray(shift, dtype=cp.float32), 0.0, 0.0
-
-    monkeypatch.setattr(
-        registration,
-        "phase_cross_correlation",
-        phase_cross_correlation,
-    )
-    image = np.zeros((63, 16, 16), dtype=np.float32)
-    image[31, 8, 8] = 1.0
-    spacing_zyx_um = (0.32, 0.098, 0.098)
-
-    transform = register_pair_to_fixed(
-        image,
-        image,
-        spacing_zyx_um=spacing_zyx_um,
-    )
-
-    recovered_shift_zyx_px = np.asarray(transform[:3, 3]) / np.asarray(spacing_zyx_um)
-    np.testing.assert_allclose(recovered_shift_zyx_px, (-0.9, 0.9, -0.8), atol=1e-5)
 
 
 def _flow_to_rgb(flow_xyz: np.ndarray, max_abs_flow: float) -> np.ndarray:
@@ -170,106 +115,6 @@ def _save_sofima_recovery_png(
     output_path.parent.mkdir(parents=True, exist_ok=True)
     fig.savefig(output_path, dpi=180)
     plt.close(fig)
-
-
-def test_register_pair_to_fixed_recovers_z_shift_for_warp_contract() -> None:
-    cp = pytest.importorskip("cupy")
-    pytest.importorskip("cucim")
-    from scipy.ndimage import shift as ndi_shift
-
-    from merfish3danalysis.utils.multiview_registration import (
-        register_pair_to_fixed,
-        warp_array_to_reference_gpu,
-    )
-
-    try:
-        cp.cuda.runtime.getDeviceCount()
-    except cp.cuda.runtime.CUDARuntimeError:
-        pytest.skip("CUDA device is not available.")
-
-    shape = (31, 64, 64)
-    z, y, x = np.indices(shape)
-    fixed = np.zeros(shape, dtype=np.float32)
-    for z0, y0, x0 in [(8, 18, 20), (16, 42, 35), (23, 30, 50)]:
-        fixed += np.exp(
-            -(((z - z0) / 2.0) ** 2 + ((y - y0) / 4.0) ** 2 + ((x - x0) / 4.0) ** 2)
-        )
-
-    true_shift_zyx_px = (2.0, 0.0, 0.0)
-    moving = ndi_shift(
-        fixed,
-        shift=true_shift_zyx_px,
-        order=1,
-        mode="constant",
-        cval=0.0,
-    ).astype(np.float32)
-    spacing_zyx_um = (0.32, 0.098, 0.098)
-    transform = register_pair_to_fixed(
-        fixed,
-        moving,
-        spacing_zyx_um=spacing_zyx_um,
-    )
-    recovered_shift_zyx_px = np.asarray(transform[:3, 3]) / np.asarray(spacing_zyx_um)
-    np.testing.assert_allclose(
-        recovered_shift_zyx_px,
-        true_shift_zyx_px,
-        atol=0.25,
-    )
-
-    warped = warp_array_to_reference_gpu(
-        moving,
-        transform_zyx_um=transform,
-        spacing_zyx_um=spacing_zyx_um,
-        reference_shape=fixed.shape,
-        gpu_id=0,
-    )
-    assert np.sqrt(np.mean((warped - fixed) ** 2)) < 1e-3
-
-
-def test_register_pair_to_fixed_residual_uses_valid_lateral_overlap() -> None:
-    cp = pytest.importorskip("cupy")
-    pytest.importorskip("cucim")
-    from scipy.ndimage import shift as ndi_shift
-
-    from merfish3danalysis.utils.multiview_registration import register_pair_to_fixed
-
-    try:
-        cp.cuda.runtime.getDeviceCount()
-    except cp.cuda.runtime.CUDARuntimeError:
-        pytest.skip("CUDA device is not available.")
-
-    shape = (24, 48, 48)
-    z, y, x = np.indices(shape)
-    fixed = np.zeros(shape, dtype=np.float32)
-    for z0, y0, x0 in [(7, 12, 14), (12, 30, 27), (18, 22, 38)]:
-        fixed += np.exp(
-            -(((z - z0) / 1.8) ** 2 + ((y - y0) / 3.2) ** 2 + ((x - x0) / 3.2) ** 2)
-        )
-
-    true_shift_zyx_px = (1.0, 8.0, -7.0)
-    moving = ndi_shift(
-        fixed,
-        shift=true_shift_zyx_px,
-        order=1,
-        mode="constant",
-        cval=0.0,
-    ).astype(np.float32)
-    spacing_zyx_um = (0.32, 0.098, 0.098)
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("error", UserWarning)
-        transform = register_pair_to_fixed(
-            fixed,
-            moving,
-            spacing_zyx_um=spacing_zyx_um,
-        )
-
-    recovered_shift_zyx_px = np.asarray(transform[:3, 3]) / np.asarray(spacing_zyx_um)
-    np.testing.assert_allclose(
-        recovered_shift_zyx_px,
-        true_shift_zyx_px,
-        atol=0.35,
-    )
 
 
 def test_sofima_estimator_recovers_object_model_warp_field(
